+++ module/actor/actor.js
@@ -14,7 +14,7 @@ export class DISActor extends Actor {
 
   /** @override */
   static async create(data, options = {}) {
-    data.token = data.token || {};
+    data.prototypeToken = data.prototypeToken || {};
     let defaults = {};
     if (data.type === "character") {
       defaults = {
@@ -35,7 +35,7 @@ export class DISActor extends Actor {
         vision: true,
       };
     }
-    mergeObject(data.token, defaults, { overwrite: false });
+    mergeObject(data.prototypeToken, defaults, { overwrite: false });
     return super.create(data, options);
   }